Combine identify and find in providers
